+++ application/internal/database/repositories/total/answerRepository.go
@@ -8,9 +8,9 @@ import (
 )
 
 type TotalRepository interface {
-	SetAnswer(ctx context.Context, testUUID, groupUUID, userUUID, quizUUID uuid.UUID, answerValue string, score int) error
+	SetAnswer(ctx context.Context, testUUID, groupUUID, userUUID, quizUUID uuid.UUID, answerValue string, score float32) error
 	Totalize(ctx context.Context, userUUID, testUUID, groupUUID uuid.UUID) error
-	AnswerScore(ctx context.Context, userUUID, testUUID, groupUUID, quizUUID uuid.UUID) (int, error)
+	AnswerScore(ctx context.Context, userUUID, testUUID, groupUUID, quizUUID uuid.UUID) (float32, error)
 	GroupTestTotals(ctx context.Context, testUUID, groupUUID uuid.UUID) ([]models.UserGroupsTests, error)
 	TestTotals(ctx context.Context, testUUID uuid.UUID) ([]models.UserGroupsTests, error)
 	AllTotals(ctx context.Context, userUUID uuid.UUID) ([]models.UserGroupsTests, error)
